feat(stella-ansi): extract the ANSI stripper into a zero-dep leaf crate
Tool output reaches the Observatory's journal route verbatim, so a
colourised child process's escapes reached a browser unstripped and
rendered as visible garbage; the clip that bounds a transcript body also
ran before the strip, so its budget could be spent on invisible bytes and
its cut could land mid-sequence. stella-observatory could not depend on
stella-tui's stripper to fix it, since that crate carries ratatui and the
observatory deliberately links almost nothing.

Move strip_ansi out of stella-tui into a new dependency-free stella-ansi
crate, the stella-diff/stella-home shape. stella-tui's ansi module
re-exports it and keeps only the ratatui-shaped emission half
(line_to_ansi, AnsiPalette). stella-observatory depends on the new crate
and strips in journal::set_journal_body, the single choke point every
journal body (text/reasoning, tool_start args, tool_result output,
step_usage's output_text, a sub_agent's report, plus system_prompt.rs and
sent_context.rs's callers) already passes through — before the clip, not
after.

Witness: two new tests in execution_routes.rs. The first seeds a
tool_result with SGR escapes and asserts the journal route's body comes
back visible-text-only; it fails on the old code because nothing on that
path ever removed an escape sequence. The second seeds a body that is
almost entirely color codes with a short visible tail, past
JOURNAL_BODY_CLIP in raw length but nowhere near it once stripped, and
asserts the row comes back whole rather than truncated — the half a
JavaScript-only fix could never reach, since the clip runs server-side.

Tests deleted: none. strip_ansi's four unit tests moved from
stella-tui/src/ansi.rs to stella-ansi/src/lib.rs unchanged.

Closes #2527
