+++ Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 resolver = "3"
 members = [
+  "crates/stella-ansi",
   "crates/stella-autonomy",
   "crates/stella-diag",
   "crates/stella-diff",
