+++ README.md
@@ -634,6 +634,7 @@ the rules it enforces, its gotchas, and the recipe for extending it.
 | [`stella-parity`](crates/stella-parity/README.md)           | The CLI-vs-API capability matrix: every capability declares a witnessed posture on both surfaces, so a feature cannot ship on one and silently miss the other |
 | [`stella-autonomy`](crates/stella-autonomy/README.md)       | The self-driving loop's decision core: AIMD controller, aperture ladder, dry-streak oracle, ledger folds — pure, shared by the CLI and the Observatory      |
 | [`stella-diff`](crates/stella-diff/README.md)               | A pure line-oriented unified diff with git's exact hunk shape and no dependencies                                                                           |
+| [`stella-ansi`](crates/stella-ansi/README.md)                | Strips ANSI codes from tool output, no dependencies                                                                                                           |
 | [`stella-embed`](crates/stella-embed/README.md)             | The embedding seam: the `Embedder` trait, the fingerprint stamped on every stored vector, and cosine ranking                                                |
 | [`stella-plugin`](crates/stella-plugin/README.md)           | Parses and validates a plugin's manifest and the wrapper socket's wire shapes, with no I/O                                                                  |
 | [`stella-transcript`](crates/stella-transcript/README.md)   | The shared transcript model plus its two renderers: HTML for the Observatory, a character grid for the terminal                                             |
