+++ crates/stella-tui/Cargo.toml
@@ -11,6 +11,10 @@ publish.workspace = true
 [dependencies]
 stella-protocol = { path = "../stella-protocol" }
 stella-tools = { path = "../stella-tools" }
+# The ANSI stripper, a leaf with no dependencies. `crate::ansi` re-exports
+# it, and keeps only the emission half (`line_to_ansi`, `AnsiPalette`) that
+# needs `ratatui`.
+stella-ansi = { path = "../stella-ansi" }
 # Only for `TerminalHold` — while this crate owns the screen, the diagnostic
 # plane's stderr sink must not write to it (see `term::TerminalGuard`).
 # stella-diag is a leaf crate by construction, so this adds no cycle.
